+++ RELEASE-NOTES.md
@@ -25,6 +25,7 @@ as release notes for end user on a Recalbox upgrade.
 ### News
 
 - Added CRT Curved shaders support
+- Added wizard to disable smooth and enable integer scale when enabling shaders
 
 ## Version 9.0.1-Pulstar